feat: only use one session and only mention the start time
Feels like doing live stream in the daytime might be challenging,
especially when someone came over. So let's just have one session in the
schedule, and if possible, doing the second session.

Also, not sure when i will finish the live stream, so to lower
expectation, let's just add start time in the schedule.

For Saturday, i want to join the OpenSUSE Bar, so the start time is a
little bit different compared to other day.

Let's see how it goes.